Add recent speaking engagements and videos
Note: This commit message is generated by ChatGPT with GPT 3.5!
Updated the content of both English and Japanese pages with the recent and upcoming speaking engagements, and added videos of previous talks. The changes include an update on the list of communities organized, as well as the addition of a page on my coffee brewing hobby.